+++ src/components/SearchFiltersMobile/SearchFiltersMobile.js
@@ -5,6 +5,8 @@ import { FormattedMessage, injectIntl, intlShape } from '../../util/reactIntl';
 import { withRouter } from 'react-router-dom';
 import omit from 'lodash/omit';
 
+import config from '../../config';
+
 import routeConfiguration from '../../routeConfiguration';
 import { parseDateFromISO8601, stringifyDateToISO8601 } from '../../util/dates';
 import { createResourceLocatorString } from '../../util/routes';
@@ -15,6 +17,7 @@ import {
   PriceFilter,
   SelectSingleFilter,
   SelectMultipleFilter,
+  SortBy,
   BookingDateRangeFilter,
 } from '../../components';
 import { propTypes } from '../../util/types';
@@ -36,6 +39,7 @@ class SearchFiltersMobileComponent extends Component {
     this.handlePrice = this.handlePrice.bind(this);
     this.handleDateRange = this.handleDateRange.bind(this);
     this.handleKeyword = this.handleKeyword.bind(this);
+    this.handleSortBy = this.handleSortBy.bind(this);
     this.initialValue = this.initialValue.bind(this);
     this.initialValues = this.initialValues.bind(this);
     this.initialPriceRangeValue = this.initialPriceRangeValue.bind(this);
@@ -129,6 +133,15 @@ class SearchFiltersMobileComponent extends Component {
     history.push(createResourceLocatorString('SearchPage', routeConfiguration(), {}, queryParams));
   }
 
+  handleSortBy(urlParam, sort) {
+    const { urlQueryParams, history } = this.props;
+    const queryParams = urlParam
+      ? { ...urlQueryParams, [urlParam]: sort }
+      : omit(urlQueryParams, urlParam);
+
+    history.push(createResourceLocatorString('SearchPage', routeConfiguration(), {}, queryParams));
+  }
+
   // Reset all filter query parameters
   resetAll(e) {
     const { urlQueryParams, history, filterParamNames } = this.props;
@@ -185,6 +198,7 @@ class SearchFiltersMobileComponent extends Component {
     const {
       rootClassName,
       className,
+      sort,
       listingsAreLoaded,
       resultsCount,
       searchInProgress,
@@ -297,6 +311,16 @@ class SearchFiltersMobileComponent extends Component {
         />
       ) : null;
 
+    const isKeywordFilterActive = !!initialKeyword;
+
+    const sortBy = config.custom.sortConfig.active ? (
+      <SortBy
+        sort={sort}
+        isKeywordFilterActive={isKeywordFilterActive}
+        onSelect={this.handleSortBy}
+      />
+    ) : null;
+
     return (
       <div className={classes}>
         <div className={css.searchResultSummary}>
@@ -334,6 +358,7 @@ class SearchFiltersMobileComponent extends Component {
               {amenitiesFilterElement}
               {priceFilterElement}
               {dateRangeFilterElement}
+              {sortBy}
             </div>
           ) : null}
 
@@ -351,6 +376,7 @@ class SearchFiltersMobileComponent extends Component {
 SearchFiltersMobileComponent.defaultProps = {
   rootClassName: null,
   className: null,
+  sort: null,
   resultsCount: null,
   searchingInProgress: false,
   selectedFiltersCount: 0,
@@ -365,6 +391,7 @@ SearchFiltersMobileComponent.propTypes = {
   rootClassName: string,
   className: string,
   urlQueryParams: object.isRequired,
+  sort: string,
   listingsAreLoaded: bool.isRequired,
   resultsCount: number,
   searchingInProgress: bool,