+++ CHANGELOG.md
@@ -14,6 +14,8 @@ way to update this template, but currently, we follow a pattern:
 
 ## Upcoming version 2020-XX-XX
 
+- [change] Move category and amenities search filters from primary filters to secondary filters. [#1275](https://github.com/sharetribe/ftw-daily/pull/1275)
+
 ## [v4.3.0] 2020-03-16
 
 - [change] Redirect user back to Stripe during Connect Onboarding Flow when user is returned to